refactor: move keyboard shortcuts functionality from store.js to util/keyModule.js
