+++ src/store.js
@@ -226,67 +226,3 @@ export function subGrid(area) {
   }
   setCurrentArea(area)
 }
-
-function enabledShortcuts(key) {
-  return ['z', 'y', 'r'].some((el) => el === key)
-}
-
-function ctrlMetaKeyHandler(event) {
-  const key = event.key.toLowerCase()
-  if (!enabledShortcuts(key)) return
-
-  event.preventDefault()
-
-  switch (key) {
-    case 'z':
-      if (event.shiftKey && canRedo) {
-        redo()
-      } else if (canUndo) {
-        undo()
-      }
-      break
-    case 'y':
-      if (canRedo) redo()
-      break
-    case 'r':
-      restart()
-      break
-  }
-}
-
-function keyHandler(event) {
-  switch (event.key) {
-    case 'Backspace':
-    case 'Delete':
-      if (currentArea.value !== mainArea.value) {
-        removeArea(currentArea.value)
-      }
-      break
-    case 'g':
-      if (currentArea.value !== mainArea.value) {
-        subGrid(currentArea.value)
-      }
-      break
-    case 'r':
-      if (currentArea?.value?.grid) {
-        addRow(currentArea.value.grid, '1fr')
-      }
-      break
-    case 'c':
-      if (currentArea?.value?.grid) {
-        addCol(currentArea.value.grid, '1fr')
-      }
-      break
-    case 'Escape':
-      if (currentArea.value !== mainArea.value) {
-        deselectCurrentArea()
-      }
-      break
-  }
-}
-
-export function keyMonitor(event) {
-  if (event.target.nodeName === 'INPUT' && event.key !== 'Escape') return
-  else if (event.ctrlKey || event.metaKey) ctrlMetaKeyHandler(event)
-  else keyHandler(event)
-}
